COC-83-fix(#79): 리팩토링에 따른 모킹 데이터와 핸들러 수정
